requirements.txt
boto3
python-dotenv
PyPDF2
mcp